Squashed commit of the following:
commit 77f4ae34c3cf1dd506bbe6bd148577c670d4ec53
Author: Chuck Jacobs <cjacobs@microsoft.com>
Date:   Tue Apr 12 20:12:52 2022 +0000

    Merged PR 2513: Removed inaccurate warp size computation for Vulkan targets

    The previous barrier optimization PR added so inaccurate code to `util::resolveWarpSize()` for Vulkan targets. This PR removes that, and fixes up some tests that depended on it.

commit 9fa84695b15c73a403649c269a4b71c31c2375d5
Author: Ritwik Das <ritdas@microsoft.com>
Date:   Tue Apr 12 17:59:15 2022 +0000

    Merged PR 2516: Add fp16 support for mfma in the DSL (+tests)

    - Add support for fp16 input and fp32 output
    - Support fp16 input and output
    - Clean up some tests

    Related work items: #3670

commit 79594afb12332e368e8f156393461838c8592a1e
Author: Ritwik Das <ritdas@microsoft.com>
Date:   Fri Apr 8 05:42:24 2022 +0000

    Merged PR 2510: Add different mfma tile sizes for FP32

    - Fix couple of offset bugs
    - Add multi-block tile sizes
    - Add unit tests

    Related work items: #3666

commit 2046b0529f9aac47c0a7ea50467a1b08a36dc5bb
Author: Mason Remy <masonr@microsoft.com>
Date:   Fri Apr 8 02:26:37 2022 +0000

    Merged PR 2511: Enable smoke test GPU matmul correctness checks

    Enable smoke test GPU matmul correctness checks

    - Also fix some FP16 scenarios
    - Add some more Accera <-> numpy mapping utilities

commit 40ad2ee4bfb1dc44ec028baeef5908204f375c30
Author: Mason Remy <masonr@microsoft.com>
Date:   Thu Apr 7 18:37:31 2022 +0000

    Merged PR 2502: Support different input array layouts for GPU caching

    Support different input array layouts for GPU caching

    This change mainly configures the thread assignments in order to get
    coalesced global memory access. The logical accessing should have
    already been correct, this is primarily for performance.

    Related work items: #3660

commit 3a62c46308be00ae70d28e3f05df9a60b4999021
Author: Chuck Jacobs <cjacobs@microsoft.com>
Date:   Thu Apr 7 16:32:34 2022 +0000

    Merged PR 2487: Barrier optimization, part 2

    This PR improves the previous barrier optimization code. It now works with non-straight-line code (if/else constructs and loops).

    It doesn't yet do the "move barriers outside of loops" optimization.

    For debugging, there's an option to output a graphviz dot file showing the graph of relevant instructions that are used during the optimization:

    ```
    acc-opt ... --barrier-opt-dot --barrier-opt-dot-filename="barrier.dot"
    ```

    Related work items: #3649

commit a306c54ce09c8257d791a7927da8ac9f80ddafa4
Author: Lisa Ong <onglisa@microsoft.com>
Date:   Thu Apr 7 00:59:58 2022 +0000

    Merged PR 2509: [nfc] sync quickstart demo from GitHub/demo branch

    Use a subset of MLAS optimizations that are sufficient to show a 3x improvement over the default schedule.

    This version was already in the Github repo for some time.

+++ accera/ir/include/exec/TensorizationInfo.h
@@ -13,7 +13,7 @@ namespace executionPlan
 {
     struct TensorizationInfo
     {
-        std::array<int, 3> dim{0,0,0};
+        std::array<int64_t, 3> dim{0,0,0};
     private:
         friend inline bool operator==(const TensorizationInfo& p1, const TensorizationInfo& p2)
         {
