+++ accera/value/include/Plan.h
@@ -219,7 +219,7 @@ namespace value
         /// <summary> Tensorize three iteration space dimensions </summary>
         /// <param name="indices"> The scalar indices to tensorize. Three indices must be specified whose dimensions must be contiguous in the iteration space dimension order. </param>
         /// <param name="numThreads"> The dimension of the tensor operation. </param>
-        void Tensorize(std::vector<ScalarIndex> indices, std::array<int, 3> dims);
+        void Tensorize(std::vector<ScalarIndex> indices, std::array<int64_t, 3> dims);
 
     private:
         friend class Schedule;
